pyproject.toml
[build-system]
requires = ["setuptools>=80", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openai-vectorstore-mcp-app"
version = "0.1.0"
description = "Authenticated file desk with a FastMCP app and companion web UI."
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "aiosqlite",
    "asyncpg",
    "colorlog",
    "debugpy",
    "fastapi",
    "fastmcp[apps]",
    "httpx",
    "itsdangerous",
    "mcp",
    "openai",
    "openai-agents",
    "openai-chatkit",
    "psycopg[binary]",
    "pydantic",
    "pydantic-settings",
    "python-multipart",
    "sqlalchemy",
    "uvicorn",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "ruff",
]

[project.scripts]
openai-vectorstore-mcp-http = "backend.http_main:main"
openai-vectorstore-mcp-stdio = "backend.main:main"

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.setuptools.packages.find]
where = ["."]
include = ["backend*"]
